Changed project structure to app directory instead of pages.
