+++ app/head.js
@@ -1,0 +1,13 @@
+export default function Head() {
+ return (
+  <>
+   <title>Plugyn Store | Safe & Cheap Metin2 Yang</title>
+
+   <meta
+    name='description'
+    content='Welcome to the plugyn yang store! Here you can get some of the cheapest yang on almost all private and official Metin2 servers. Contact us today for a custom offer!'
+   />
+   <link rel='icon' href='/favicon.ico' />
+  </>
+ );
+}
