+++ app/layout.tsx
@@ -1,0 +1,13 @@
+export default function RootLayout({
+ // Layouts must accept a children prop.
+ // This will be populated with nested layouts or pages
+ children,
+}: {
+ children: React.ReactNode;
+}) {
+ return (
+  <html lang='en'>
+   <body>{children}</body>
+  </html>
+ );
+}
