+++ next.config.js
@@ -1,7 +1,10 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  reactStrictMode: true,
-  swcMinify: true,
-}
+ reactStrictMode: true,
+ swcMinify: true,
+ experimental: {
+  appDir: true,
+ },
+};
 
-module.exports = nextConfig
+module.exports = nextConfig;
